+++ inference-engine/src/snippets/src/itt.hpp
@@ -9,8 +9,7 @@
 
 #pragma once
 
-#include <openvino/cc/selective_build.h>
-#include <openvino/itt.hpp>
+#include <openvino/cc/ngraph/itt.hpp>
 
 namespace ngraph {
 namespace pass {
@@ -31,29 +30,15 @@ OV_CC_DOMAINS(internal_op);
  * INTERNAL_OP_SCOPE macro allows to disable parts of internal nGraph operations if they are not used
  */
 #if defined(SELECTIVE_BUILD_ANALYZER)
-#define RUN_ON_FUNCTION_SCOPE(region) OV_SCOPE(ngraph_pass, OV_PP_CAT(region, _run_on_function))
-#define MATCHER_SCOPE(region)                                                                   \
-    const std::string matcher_name(OV_PP_TOSTRING(region))
 
 #define INTERNAL_OP_SCOPE(region) OV_SCOPE(internal_op, region)
 
 #elif defined(SELECTIVE_BUILD)
 
-#define MATCHER_SCOPE_(scope, region)                                                           \
-    if (OV_CC_SCOPE_IS_ENABLED(OV_PP_CAT3(scope, _, region)) == 0)                              \
-    throw ngraph::ngraph_error(std::string(OV_PP_TOSTRING(OV_PP_CAT3(scope, _, region))) +      \
-                               " is disabled!")
-
-#define MATCHER_SCOPE(region)                                                                   \
-    const std::string matcher_name(OV_PP_TOSTRING(region));                                           \
-    if (OV_CC_SCOPE_IS_ENABLED(OV_PP_CAT3(ngraph_pass, _, region)) == 0)                        \
-        return
 #define INTERNAL_OP_SCOPE(region) MATCHER_SCOPE_(internal_op, region)
-#define RUN_ON_FUNCTION_SCOPE(region) MATCHER_SCOPE_(ngraph_pass, OV_PP_CAT(region, _run_on_function))
 
 #else
-#define MATCHER_SCOPE(region)                                                                   \
-    const std::string matcher_name(OV_PP_TOSTRING(region))
+
 #define INTERNAL_OP_SCOPE(region)
-#define RUN_ON_FUNCTION_SCOPE(region)
+
 #endif
